requirements-minimal.txt
fastapi==0.116.1
uvicorn==0.35.0
xgboost==3.0.2
pandas==2.3.1
scikit-learn==1.7.1
pydantic==2.11.7
google-cloud-storage==2.18.0
python-dotenv==1.0.1
